serverless/aws/requirements.txt
aioboto3==10.1.0
aiobotocore==2.4.0
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
boto3==1.24.59
botocore==1.27.59
charset-normalizer==2.1.1
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.1
frozenlist==1.3.1
idna==3.4
jmespath==1.0.1
multidict==6.0.2
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
yarl==1.8.1
